Only delete Discord webhooks when they are the entire message content (#79)

+++ src/bot/exts/filters/webhook_remover.py
@@ -9,7 +9,7 @@ from discord import Colour, Message, NotFound
 from discord.ext.commands import Cog
 
 from bot.bot import Bot
-from bot.constants import Channels, Colours, Icons
+from bot.constants import Channels, Colours, Icons, Roles
 from bot.exts.core.log import Log
 from bot.utils.messages import format_user
 
@@ -51,6 +51,18 @@ class WebhookRemover(Cog):
         async with self.bot.http_session.delete(webhook_url) as response:
             # The Discord API Returns a 204 NO CONTENT response on success.
             deleted_successfully = response.status == 204
+
+        # The webhook should only be actioned if it is the only content of a message,
+        # and within the defined exemption parameters.
+        user_roles = [role.id for role in message.author.roles]
+        if all(
+            [
+                message.content != webhook_url,
+                message.channel.category.id == Channels.soc_category,
+                Roles.security in user_roles,
+            ]
+        ):
+            return
 
         try:
             await message.delete()
